+++ test/CMakeLists.txt
@@ -356,6 +356,7 @@ make_test(SRC pad_borders.cpp CXX11)
 make_test(SRC pinverse.cpp SERIAL)
 make_test(SRC qr_dense.cpp SERIAL)
 make_test(SRC random.cpp)
+make_test(SRC rng_quality.cpp BACKENDS "cuda;opencl" SERIAL)
 make_test(SRC range.cpp)
 make_test(SRC rank_dense.cpp SERIAL)
 make_test(SRC reduce.cpp CXX11)
